crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.13.0
scrapy==2.11.0
pymongo==4.5.0
redis==5.0.0
python-dotenv==1.0.0
aiohttp==3.9.0
Pillow==10.0.0
